requirements.txt
Django==3.1.2
djangorestframework==3.12.4
Pillow==8.0.1
psycopg2==2.8.6
virtualenv==20.1.0
whitenoise==5.2.0
